requirements.txt
aiohttp
typer[all]
PyYAML
tqdm
rich